app/utils/jwt_handler.py

from jose import jwt
from datetime import datetime, timedelta
from app.config import JWT_SECRET

def create_token(data: dict, expires_delta=3600):
    to_encode = data.copy()
    expire = datetime.utcnow() + timedelta(seconds=expires_delta)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, JWT_SECRET, algorithm="HS256")

def decode_token(token: str):
    return jwt.decode(token, JWT_SECRET, algorithms=["HS256"])
