requirements.txt

fastapi
uvicorn
motor
python-multipart
python-jose
passlib[bcrypt]
email-validator
cryptography
